requirements.txt
APScheduler==3.7.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.5
channels==2.1.7
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.4.6
daphne==2.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.10
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-formtools==2.2
django-import-export==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
ecdsa==0.16.1
et-xmlfile==1.0.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
lxml==4.5.0
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.3
Naked==0.1.31
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.9
paramiko==1.18.5
Pillow==7.1.2
pluggy==0.13.1
psutil==5.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyHamcrest==2.0.2
PyJWT==2.0.0
PyMySQL==1.0.2
PyNaCl==1.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-runner==5.2
python-dateutil==2.5.3
python-docx==0.8.7
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.23.0
requests-toolbelt==0.9.1
service-identity==18.1.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
Twisted==18.7.0
txaio==21.2.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
